requirements.txt
streamlit>=1.25.0
python-dotenv
groq>=0.0.23
sqlite3